LayoutTests/platform/chromium-win/fast/js/function-prototype-descriptor-expected.txt

Test for function.prototype's property descriptor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS descriptor['writable'] is true
FAIL descriptor['enumerable'] should be true. Was false.
PASS descriptor['configurable'] is false
PASS descriptor['writable'] is true
FAIL descriptor['enumerable'] should be true. Was false.
PASS descriptor['configurable'] is false
PASS Object.defineProperty(c, 'prototype', { get: function(){} }) threw exception TypeError: Cannot redefine property: defineProperty.
PASS descriptor['writable'] is true
FAIL descriptor['enumerable'] should be true. Was false.
PASS descriptor['configurable'] is false
PASS successfullyParsed is true

TEST COMPLETE
